% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBS.R
\name{DWBS}
\alias{DWBS}
\title{Find Changepoints Using Depth-Based Wild Binary Segmentation}
\usage{
DWBS(data, numInt = 10, thresh = 1.3584, depth = "hs")
}
\arguments{
\item{numInt}{Number of subintervals to be generated}

\item{depth}{Depth function of choice}
}
\value{
A list of changepoints
}
\description{
Find Changepoints Using Depth-Based Wild Binary Segmentation
}
