% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{epidemic_test}
\alias{epidemic_test}
\title{Test for an epidemic period in data}
\usage{
epidemic_test(data, ranks = NULL, depth = c("FM", "RPD", "LTR", "FMd", "RPDd"))
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{ranks}{Optional if data is already ranked.}

\item{depth}{Depth function of choice.}
}
\value{
Estimated start and end of epidemic period with p-value.
}
\description{
Test for an epidemic period in data
}
\note{
The \code{depth} arguments are as follows:
\itemize{
\item \code{FM}: Frainman-Muniz depth
\item \code{RPD}: Random projection depth
\item \code{LTR}: \eqn{L^2} norm depth, most suitable for detecting changes in the norm
\item \code{FMd}: Frainman-Muniz depth of the data and its first order derivative
\item \code{RPDd}: Random projection depth of the data and its first order derivative
}
}
\examples{
epi_test <- rbind(replicate(3,rnorm(200)),
                  replicate(3,rnorm(200,10)),
                  replicate(3,rnorm(200,0.2)))

epidemic_test(epi_test)

}
