% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRI.R
\name{get_4D_changepoints}
\alias{get_4D_changepoints}
\title{Detect changepoints in a resting state fMRI scan}
\usage{
get_4D_changepoints(data, p = 100, k = 0.3)
}
\arguments{
\item{data}{A four dimensional array, where the fourth dimension is each time stamp.}

\item{p}{Number of random vector projections, set to 100 by default}

\item{k}{Part of penalty constant passed to pruned exact linear time algorithm.}
}
\value{
A list of changepoints
}
\description{
Detect changepoints in a resting state fMRI scan
}
\note{
The penalty is of the form \deqn{3.74 + k*\sqrt{n}} where \eqn{n} is the
number of observations. In the case that there is potentially correlated
observations, the parameter could be set to \eqn{k=1}. More information could
be found in the reference.
}
\references{
Ramsay, K., & Chenouri, S. (2025). Robust changepoint detection
  in the variability of multivariate functional data. Journal of
  Nonparametric Statistics. https://doi.org/10.1080/10485252.2025.2503891
}
