% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KWCChangepoint-package.R
\docType{package}
\name{KWCChangepoint-package}
\alias{KWCChangepoint}
\alias{KWCChangepoint-package}
\title{KWCChangepoint: User-friendly Functional and Multivariate Changepoint Detection}
\description{
The package mainly focuses on changepoint detection in the context of multivariate or functional data.
}
\section{Links}{

\itemize{
\item GitHub: \url{https://github.com/adeeb99/KWCChangepoint}
\item Bug reports: \url{https://github.com/adeeb99/KWCChangepoint/issues}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adeeb99/KWCChangepoint}
  \item Report bugs at \url{https://github.com/adeeb99/KWCChangepoint/issues}
}

}
\author{
\strong{Maintainer}: Adeeb Rouhani \email{adeeb.rouhani@gmail.com} (\href{https://orcid.org/0009-0007-5294-2226}{ORCID})

Authors:
\itemize{
  \item Kelly Ramsay (\href{https://orcid.org/0000-0001-5615-2052}{ORCID})
}

}
