% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{uni_scale}
\alias{uni_scale}
\title{Find changepoints in univariate data based on scale}
\usage{
uni_scale(data, beta = 10)
}
\arguments{
\item{data}{A vector or one-dimensional array.}

\item{beta}{Numeric penalty constant passed to pruned exact linear time
algorithm, 10 by default.}
}
\value{
List of changepoints.
}
\description{
Find changepoints in univariate data based on scale
}
\examples{
scale_test <- c(rnorm(100, sd=5), # before change in sale
                rnorm(100, sd=1)) # after change in scale
uni_scale(scale_test)


}
