% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{uniMean}
\alias{uniMean}
\title{Find changepoint in univariate data based on mean}
\usage{
uniMean(data, beta = 10)
}
\arguments{
\item{data}{Univariate data}

\item{beta}{Numeric penalty constant passed to PELT}
}
\value{
List of changepoints
}
\description{
Find changepoint in univariate data based on mean
}
