% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{uniScale}
\alias{uniScale}
\title{Find Changepoints in Univariate Data Based on Scale}
\usage{
uniScale(data, beta = 10)
}
\arguments{
\item{data}{Univariate data}

\item{beta}{Numeric penalty constant passed to PELT, 10 by default}
}
\value{
List of changepoints
}
\description{
Find Changepoints in Univariate Data Based on Scale
}
