% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{AMOC_test}
\alias{AMOC_test}
\title{Conduct an AMOC (at most one changepoint) hypothesis test}
\usage{
AMOC_test(data, ranks = NULL, useRank = FALSE, depth = "spat", boundary = 1)
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{ranks}{Optional if data is already ranked}

\item{useRank}{FALSE by defalut, set to TRUE to use your ranks}

\item{depth}{Depth function of choice. It is 'spat' for spatial depth by
default. User can also choose 'mahal' for Mahalanobis, 'mahal75' for
Mahalanobis MCD, or 'hs' for halfspace depth.}

\item{boundary}{}
}
\value{
An estimated changepoint with a p-value.
}
\description{
Conduct an AMOC (at most one changepoint) hypothesis test
}
