% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{amoc_test}
\alias{amoc_test}
\title{Conduct an AMOC (at most one changepoint) hypothesis test}
\usage{
amoc_test(data, ranks = NULL, depth = c("spat", "hs", "mahal", "mahal75"))
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{ranks}{Optional if data is already ranked}

\item{depth}{Depth function of choice. It is 'spat' for spatial depth by
default. User can also choose 'mahal' for Mahalanobis, 'mahal75' for
Mahalanobis MCD, or 'hs' for halfspace depth.}
}
\value{
An estimated changepoint with a p-value.
}
\description{
Conduct an AMOC (at most one changepoint) hypothesis test
}
\examples{
test_data <- rbind(replicate(3,rnorm(200,1,1)), #before changepoint
                   replicate(3,rnorm(200,1,5))) #after changepoint

amoc_test(test_data)
}
