% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{amoc_test}
\alias{amoc_test}
\title{Conduct an AMOC (at most one changepoint) hypothesis test}
\usage{
amoc_test(data, ranks = NULL, depth = c("FM", "RPD", "LTR", "FMd", "RPDd"))
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{ranks}{Optional if data is already ranked.}

\item{depth}{Depth function of choice.}
}
\value{
An estimated changepoint with a p-value.
}
\description{
Conduct an AMOC (at most one changepoint) hypothesis test
}
\note{
The \code{depth} arguments are as follows:
\itemize{
\item \code{FM}: Frainman-Muniz depth
\item \code{RPD}: Random projection depth
\item \code{LTR}: \eqn{L^2} norm depth, most suitable for detecting changes in the norm
\item \code{FMd}: Frainman-Muniz depth of the data and its first order derivative
\item \code{RPDd}: Random projection depth of the data and its first order derivative
}
}
\examples{
test_data <- rbind(replicate(3,rnorm(200,1,1)), #before changepoint
                   replicate(3,rnorm(200,1,5))) #after changepoint

amoc_test(test_data)
}
