% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{AMOC_test}
\alias{AMOC_test}
\title{Conduct an AMOC (at most one changepoint) hypothesis test}
\usage{
AMOC_test(data, ranks = NULL, useRank = FALSE, setDepth = "spat", boundary = 1)
}
\arguments{
\item{data}{Data in matrix form}

\item{ranks}{Optional if data is already ranked}

\item{useRank}{FALSE by defalut, set to TRUE to use your ranks}

\item{setDepth}{Depth function of choice}

\item{boundary}{}
}
\value{
An estimated changepoint with a p-value
}
\description{
Conduct an AMOC (at most one changepoint) hypothesis test
}
