% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBS.R
\name{DWBS}
\alias{DWBS}
\title{Find changepoints using depth-based wild binary segmentation}
\usage{
DWBS(data, N, d, numInt = 10, thresh = 1.3584, depth = "hs")
}
\arguments{
\item{numInt}{Number of subintervals to be generated}

\item{depth}{Depth function of choice}
}
\value{
A list of changepoints
}
\description{
Find changepoints using depth-based wild binary segmentation
}
