% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBS.R
\name{dwbs}
\alias{dwbs}
\title{Find changepoints using data-driven depth-based wild binary segmentation}
\usage{
dwbs(
  data,
  numInt = 10,
  thresh = 1.3584,
  alpha = 1,
  depth = c("hs", "spat", "mahal", "mahal75")
)
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{numInt}{Number of subintervals to be generated.}

\item{thresh}{Numeric scalar; detection threshold. Larger values make
detection more conservative.}

\item{alpha}{Set as 1 by default, applying a standard SIC penalty. Set to a
number larger than 1 for a strengthened SIC.}

\item{depth}{Depth function of choice, 'hs' for halfspace depth by
default. Users can also choose 'mahal' for Mahalanobis, 'mahal75' for
Mahalanobis MCD, 'spat' for spatial depth.}
}
\value{
A list of changepoints.
}
\description{
Find changepoints using data-driven depth-based wild binary segmentation
}
\examples{
exdata <- rbind(replicate(3,rnorm(200)),
                replicate(3,rnorm(200,10)),
                replicate(3,rnorm(200,0.2)))
dwbs(data = exdata)

# Increasing `numInt` will result in more accurate detection
dwbs(data = exdata, numInt = 100)

}
\references{
Killick, R., P. Fearnhead, and I. A. Eckley. “Optimal Detection
of Changepoints With a Linear Computational Cost.” Journal of the American
Statistical Association 107, no. 500 (2012): 1590–98.
https://doi.org/10.1080/01621459.2012.737745.

Ramsay, K., & Chenouri, S. (2023). Robust nonparametric multiple
changepoint detection for multivariate variability. Econometrics and
Statistics. https://doi.org/10.1016/j.ecosta.2023.09.001
}
