% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWBS.R
\name{DWBS}
\alias{DWBS}
\title{Find changepoints using depth-based wild binary segmentation}
\usage{
DWBS(data, numInt = 10, thresh = 1.3584, depth = "hs")
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{numInt}{Number of subintervals to be generated.}

\item{thresh}{Numeric scalar; detection threshold. Larger values make
detection more conservative.}

\item{depth}{Depth function of choice. It is 'hs' for halfspace depth by
default. User can also choose 'mahal' for Mahalanobis, 'mahal75' for
Mahalanobis MCD, 'spat' for spatial depth.}
}
\value{
A list of changepoints.
}
\description{
Find changepoints using depth-based wild binary segmentation
}
\references{
Killick, R., P. Fearnhead, and I. A. Eckley. “Optimal Detection
of Changepoints With a Linear Computational Cost.” Journal of the American
Statistical Association 107, no. 500 (2012): 1590–98.
https://doi.org/10.1080/01621459.2012.737745.
}
