% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{Epidemic_test}
\alias{Epidemic_test}
\title{Test For an Epidemic Period in Data}
\usage{
Epidemic_test(data, ranks = NULL, useRank = FALSE, depth = "spat")
}
\arguments{
\item{data}{Matrix of data}

\item{ranks}{Optional if data is already ranked}

\item{useRank}{FALSE by defalut, set to TRUE to use your ranks}

\item{depth}{Depth function of choice}
}
\value{
Estimated start and end of epidemic period with p-value
}
\description{
Test For an Epidemic Period in Data
}
