% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{Epidemic_test}
\alias{Epidemic_test}
\title{Test For an Epidemic Period in Data}
\usage{
Epidemic_test(data, ranks = NULL, useRank = FALSE, depth = "spat")
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{ranks}{Optional if data is already ranked}

\item{useRank}{FALSE by defalut, set to TRUE to use your ranks}

\item{depth}{Depth function of choice. It is 'spat' for spatial depth by
default. User can also choose 'mahal' for Mahalanobis, 'mahal75' for
Mahalanobis MCD, or 'hs' for halfspace depth.}
}
\value{
Estimated start and end of epidemic period with p-value
}
\description{
Test For an Epidemic Period in Data
}
