% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{epidemic_test}
\alias{epidemic_test}
\title{Test for an epidemic period in data}
\usage{
epidemic_test(data, ranks = NULL, depth = c("spat", "hs", "mahal", "mahal75"))
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{ranks}{Optional if data is already ranked.}

\item{depth}{Depth function of choice. It is 'spat' for spatial depth by
default. User can also choose 'mahal' for Mahalanobis, 'mahal75' for
Mahalanobis MCD, or 'hs' for halfspace depth.}
}
\value{
Estimated start and end of epidemic period with p-value.
}
\description{
Test for an epidemic period in data
}
\examples{
epi_test <- rbind(replicate(3,rnorm(200)),
                  replicate(3,rnorm(200,10)),
                  replicate(3,rnorm(200,0.2)))

epidemic_test(epi_test)

}
