% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FKWC.R
\name{FKWC}
\alias{FKWC}
\title{Find Changepoints Using Functional Kruskall-Wallis Tests for Covariance
Algorithm}
\usage{
FKWC(funcdata, depth = "FM_depth", k = 0.25)
}
\arguments{
\item{funcdata}{Functional data in fdata form, where each row is an
observation and each column is a dimension.}

\item{depth}{Depth function of choice}

\item{k}{Part of penalty constant passed to PELT}
}
\value{
A list of changepoints
}
\description{
Find Changepoints Using Functional Kruskall-Wallis Tests for Covariance
Algorithm
}
