% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FKWC.R
\name{FKWC_multisample}
\alias{FKWC_multisample}
\title{Multisample hypothesis test for differences in covariance operators using
Functional Kruskal–Wallis Tests for Covariance.}
\usage{
FKWC_multisample(data, derivs, g, p = 20)
}
\arguments{
\item{data}{Functional data in fdata form, where each row is an observation
and each column is a dimension.}

\item{derivs}{First order derivative of the functional data in fdata form.}

\item{g}{Factor object that indicates which sample each row of data belongs
to.}

\item{p}{Number of random projections to be generated in order to compute
random projection depths of the data.}
}
\value{
A list containing the test statistic and the p-value.
}
\description{
Multisample hypothesis test for differences in covariance operators using
Functional Kruskal–Wallis Tests for Covariance.
}
\references{
Ramsay, K., & Chenouri, S. (2024). Robust nonparametric
  hypothesis tests for differences in the covariance structure of functional
  data. Canadian Journal of Statistics, 52 (1), 43–78.
  https://doi.org/10.1002/cjs.11767
}
