% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRI.R
\name{get_4D_changepoints}
\alias{get_4D_changepoints}
\title{Detect changepoints in an fMRI}
\usage{
get_4D_changepoints(imag_array, M = 100, C1 = 0.3, C2 = 3.74)
}
\arguments{
\item{imag_array}{A 4D object}

\item{C2}{}
}
\value{
A list of changepoints
}
\description{
Detect changepoints in an fMRI
}
