% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{getRanks}
\alias{getRanks}
\title{Rank Multivariate Data on Depth Values}
\usage{
getRanks(data, depth = "spat")
}
\arguments{
\item{data}{Matrix of data}

\item{depth}{Depth function of choice}
}
\value{
A list of ranks for each observation
}
\description{
Rank Multivariate Data on Depth Values
}
