% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KWCChangepoint-package.R
\docType{package}
\name{KWCChangepoint-package}
\alias{KWCChangepoint}
\alias{KWCChangepoint-package}
\title{KWCChangepoint: User-friendly Functional and Multivariate Changepoint Detection}
\description{
A group of functions that detect and test for changes in covariance structures of functional data, as well as changepoint detection for multivariate data more generally. Method for detecting non-stationarity in resting state fMRI scans as seen in Ramsay, K., & Chenouri, S. (2025) \doi{10.1080/10485252.2025.2503891} is implemented in 'fmri_changepoints()'. Also includes depth- and rank-based implementation of the WBS algorithm for detecting multiple changepoints in multivariate data.
}
\section{Links}{

\itemize{
\item GitHub: \url{https://github.com/adeeb99/KWCChangepoint}
\item Bug reports: \url{https://github.com/adeeb99/KWCChangepoint/issues}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adeeb99/KWCChangepoint}
  \item \url{https://adeeb99.github.io/KWCChangepoint/}
  \item Report bugs at \url{https://github.com/adeeb99/KWCChangepoint/issues}
}

}
\author{
\strong{Maintainer}: Adeeb Rouhani \email{adeeb.rouhani@gmail.com} (\href{https://orcid.org/0009-0007-5294-2226}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kelly Ramsay (\href{https://orcid.org/0000-0001-5615-2052}{ORCID})
}

}
