% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKWP.R
\name{mkwp}
\alias{mkwp}
\title{Find changepoints in multivariate data using multivariate Kruskal-Wallis
PELT.}
\usage{
mkwp(data, depth = c("mahal", "mahal75", "spat", "hs"), k = 0.2)
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{depth}{Depth function of choice. It is 'mahal' for Mahalanobis depth by
default. User can also choose 'mahal75' for Mahalanobis MCD, 'hs' for
halfspace, or 'spat' for spatial depth.}

\item{k}{Penalty constant passed to pruned exact linear time algorithm.}
}
\value{
A list of changepoints.
}
\description{
Find changepoints in multivariate data using multivariate Kruskal-Wallis
PELT.
}
\examples{
set.seed(111)
multi_data <-rbind(replicate(3,rnorm(200)),
                   replicate(3,rnorm(200,10)),
                   replicate(3,rnorm(200,0.2)))
mkwp(multi_data)

}
\references{
Ramsay, K., & Chenouri, S. (2023). Robust nonparametric multiple
changepoint detection for multivariate variability. Econometrics and
Statistics. https://doi.org/10.1016/j.ecosta.2023.09.001
}
