% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKWP.R
\name{MKWP}
\alias{MKWP}
\title{Find Changepoints Using Multivariate Kruskal-Wallis PELT}
\usage{
MKWP(data, depth = "mahal", beta = 10, custom_depth_function = NULL)
}
\arguments{
\item{data}{Matrix of data}

\item{depth}{Depth function of choice}

\item{beta}{Numeric penalty constant passed to PELT}

\item{custom_depth_function}{Use your own custom depth function (NULL by default)}
}
\value{
A list of changepoints
}
\description{
Find Changepoints Using Multivariate Kruskal-Wallis PELT
}
