% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKWP.R
\name{MKWP}
\alias{MKWP}
\title{Find changepoints in multivariate data using multivariate Kruskal-Wallis
PELT.}
\usage{
MKWP(data, depth = "mahal", k = 0.2, custom_depth_function = NULL)
}
\arguments{
\item{data}{A matrix or dataframe, where each row is an observation and each
column is a dimension.}

\item{depth}{Depth function of choice. It is 'mahal' for Mahalanobis depth by
default. User can also choose 'mahal75' for Mahalanobis MCD, 'hs' for
halfspace, or 'spat' for spatial depth.}

\item{k}{Part of numeric penalty constant passed to PELT}

\item{custom_depth_function}{Use your own custom depth function (NULL by
default)}
}
\value{
A list of changepoints
}
\description{
Find changepoints in multivariate data using multivariate Kruskal-Wallis
PELT.
}
\references{
Ramsay, K., & Chenouri, S. (2023). Robust nonparametric multiple
changepoint detection for multivariate variability. Econometrics and
Statistics. https://doi.org/10.1016/j.ecosta.2023.09.001
}
