% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{uni_mean}
\alias{uni_mean}
\title{Find changepoint in univariate data based on mean}
\usage{
uni_mean(data, beta = 10)
}
\arguments{
\item{data}{A vector or one-dimensional array.}

\item{beta}{Numeric penalty constant passed to pruned exact linear time
algorithm.}
}
\value{
List of changepoints.
}
\description{
Find changepoint in univariate data based on mean
}
\examples{
mean_test <- c(rnorm(100, mean = 0), # before change in mean
               rnorm(100, mean = 5)) # after change in mean
uni_mean(mean_test)


}
\references{
Killick, R., P. Fearnhead, and I. A. Eckley. “Optimal Detection
of Changepoints With a Linear Computational Cost.” Journal of the American
Statistical Association 107, no. 500 (2012): 1590–98.
https://doi.org/10.1080/01621459.2012.737745.
}
