% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{uniScale}
\alias{uniScale}
\title{Find changepoints in univariate data based on scale}
\usage{
uniScale(data, beta = 10)
}
\arguments{
\item{data}{Univariate data.}

\item{beta}{Numeric penalty constant passed to pruned exact linear time
algorithm, 10 by default.}
}
\value{
List of changepoints.
}
\description{
Find changepoints in univariate data based on scale
}
